+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Pre-commit hook now validates for unused dependencies in mix.lock
+
 ## [0.1.0] - 2025-07-25
 
 ### Added